cid入力と最近プレイした譜面リストをトップページに移動
