+++ frontend/app/[locale]/main/play/clientPage.tsx
@@ -29,117 +29,25 @@ import { ShareBox } from "@/share/placeholder/shareBox.js";
 import { titleShare, titleWithSiteName } from "@/common/title.js";
 import * as v from "valibot";
 import { SlimeSVG } from "@/common/slime.js";
+import { useShareModal } from "./shareModal.jsx";
 
 export default function PlayTab({ locale }: { locale: string }) {
   const t = useTranslations("main.play");
-  const th = useTranslations("share");
   const te = useTranslations("error");
   const { isMobileMain } = useDisplayMode();
 
-  const [recentBrief, setRecentBrief] = useState<ChartLineBrief[]>();
-  const [fetchRecentAll, setFetchRecentAll] = useState<boolean>(false);
   const [popularBrief, setPopularBrief] = useState<
     ChartLineBrief[] | "error" | undefined
   >(undefined);
-  const [fetchPopularAll, setFetchPopularAll] = useState<boolean>(false);
   const [latestBrief, setLatestBrief] = useState<
     ChartLineBrief[] | "error" | undefined
   >(undefined);
-  const [fetchLatestAll, setFetchLatestAll] = useState<boolean>(false);
-  const [originalBrief, setOriginalBrief] = useState<ChartLineBrief[]>();
-  const fetchOriginalAll = true;
-  const [sampleBrief, setSampleBrief] = useState<ChartLineBrief[]>();
-  const fetchSampleAll = true;
+  const [originalBrief, setOriginalBrief] = useState<ChartLineBrief[]>(originalCId.map((cid) => ({ cid, fetched: false, original: true })));
+  const [sampleBrief, setSampleBrief] = useState<ChartLineBrief[]>(sampleCId.map((cid) => ({ cid, fetched: false })));
 
-  // exclusiveをセット(指定したもの以外を非表示にする) → 200ms後、showAllをセット(指定したものの内容を全て表示する)
-  // mobileではごちゃごちゃやってもスクロールが入るせいできれいに見えないので瞬時に切り替える
-  const [showExclusiveMode, setShowExclusiveMode] = useState<
-    null | "recent" | "popular" | "latest"
-  >(null);
-  const [showAllMode, setShowAllMode] = useState<
-    null | "recent" | "popular" | "latest"
-  >(null);
-  const goExclusiveMode = useCallback(
-    (mode: "recent" | "popular" | "latest") => {
-      window.history.replaceState(null, "", "#"); // これがないとなぜか #recent から元のページにブラウザバックできなくなる場合があるけどなぜ?
-      window.history.pushState(null, "", "#" + mode);
-      setShowExclusiveMode(mode);
-      if (isMobileMain) {
-        setShowAllMode(mode);
-        window.scrollTo(0, 0);
-      } else {
-        setTimeout(() => setShowAllMode(mode), 200);
-      }
-    },
-    [isMobileMain],
-  );
-
-  const [modalCId, setModalCId] = useState<string | null>(null);
-  const [modalBrief, setModalBrief] = useState<ChartBrief | null>(null);
-  const [modalRecord, setModalRecord] = useState<RecordGetSummary[]>([]);
-  const [modalAppearing, setModalAppearing] = useState<boolean>(false);
-  const openModal = useCallback(
-    (cid: string, brief: ChartBrief | undefined) => {
-      if (brief) {
-        if (window.location.pathname !== `/share/${cid}`) {
-          // pushStateではpopstateイベントは発生しない
-          window.history.pushState(null, "", `/share/${cid}`);
-        }
-        setModalCId(cid);
-        setModalBrief(brief);
-        fetch(process.env.BACKEND_PREFIX + `/api/record/${cid}`)
-          .then((res) => {
-            if (res.ok) {
-              return res.json();
-            } else {
-              throw new Error("failed to fetch record");
-            }
-          })
-          .then((record) => setModalRecord(record))
-          .catch(() => setModalRecord([]));
-        document.title = titleShare(th, cid, brief);
-        setTimeout(() => setModalAppearing(true));
-      }
-    },
-    [th],
-  );
-
-  // modalのcloseと、exclusiveModeのリセットは window.history.back(); でpopstateイベントを呼び出しその中で行われる
-  useEffect(() => {
-    const handler = () => {
-      if (window.location.pathname.startsWith("/share/")) {
-        const cid = window.location.pathname.slice(7);
-        fetchBrief(cid).then((res) => {
-          openModal(cid, res.brief);
-        });
-      } else {
-        setModalAppearing(false);
-        document.title = titleWithSiteName(t("title"));
-        setTimeout(() => {
-          setModalCId(null);
-          setModalBrief(null);
-        }, 200);
-        if (window.location.hash.length >= 2) {
-          goExclusiveMode(
-            window.location.hash.slice(1) as "recent" | "popular" | "latest",
-          );
-        } else {
-          setShowAllMode(null);
-          setShowExclusiveMode(null);
-        }
-      }
-    };
-    window.addEventListener("popstate", handler);
-    return () => window.removeEventListener("popstate", handler);
-  }, [goExclusiveMode, openModal, t]);
+  const {modal, openModal} = useShareModal(locale);
 
   useEffect(() => {
-    const recentCId = getRecent("play").reverse();
-    setRecentBrief(recentCId.map((cid) => ({ cid, fetched: false })));
-    setOriginalBrief(
-      originalCId.map((cid) => ({ cid, fetched: false, original: true })),
-    );
-    setSampleBrief(sampleCId.map((cid) => ({ cid, fetched: false })));
     void (async () => {
       try {
         const latestRes = await fetch(
@@ -183,211 +91,64 @@ export default function PlayTab({ locale }: { locale: string }) {
   }, []);
   useEffect(() => {
     void (async () => {
-      if (recentBrief) {
-        const { changed, briefs } = await fetchAndFilterBriefs(
-          recentBrief,
-          fetchRecentAll,
-        );
-        if (changed) {
-          setRecentBrief(briefs);
-          updateRecent("play", briefs.map(({ cid }) => cid).reverse());
-        }
-      }
-    })();
-  }, [recentBrief, fetchRecentAll]);
-  useEffect(() => {
-    void (async () => {
       if (Array.isArray(latestBrief)) {
         const { changed, briefs } = await fetchAndFilterBriefs(
           latestBrief,
-          fetchLatestAll,
+          false,
         );
         if (changed) {
           setLatestBrief(briefs);
         }
       }
     })();
-  }, [latestBrief, fetchLatestAll]);
+  }, [latestBrief]);
   useEffect(() => {
     void (async () => {
       if (Array.isArray(popularBrief)) {
         const { changed, briefs } = await fetchAndFilterBriefs(
           popularBrief,
-          fetchPopularAll,
+          false,
         );
         if (changed) {
           setPopularBrief(briefs);
         }
       }
     })();
-  }, [popularBrief, fetchPopularAll]);
+  }, [popularBrief]);
   useEffect(() => {
     void (async () => {
       if (originalBrief) {
         const { changed, briefs } = await fetchAndFilterBriefs(
           originalBrief,
-          fetchOriginalAll,
+          true,
         );
         if (changed) {
           setOriginalBrief(briefs);
         }
       }
     })();
-  }, [originalBrief, fetchOriginalAll]);
+  }, [originalBrief]);
   useEffect(() => {
     void (async () => {
       if (sampleBrief) {
         const { changed, briefs } = await fetchAndFilterBriefs(
           sampleBrief,
-          fetchSampleAll,
+          true,
         );
         if (changed) {
           setSampleBrief(briefs);
         }
       }
     })();
-  }, [sampleBrief, fetchSampleAll]);
-
-  const [cidErrorMsg, setCIdErrorMsg] = useState<string>("");
-  const [cidFetching, setCidFetching] = useState<boolean>(false);
-  const gotoCId = async (cid: string) => {
-    setCIdErrorMsg("");
-    setCidFetching(true);
-    const res = await fetchBrief(cid, true);
-    if (res.ok) {
-      // router.push(`/share/${cid}`);
-      openModal(cid, res.brief);
-    } else {
-      if (res.is404) {
-        setCIdErrorMsg(te("api.chartIdNotFound"));
-      } else {
-        setCIdErrorMsg(te("unknownApiError"));
-      }
-    }
-    setCidFetching(false);
-  };
+  }, [sampleBrief]);
 
   return (
     <IndexMain
-      tab={1}
+      title={t("title")}
+      tabKey="play"
       locale={locale}
-      modal={
-        modalCId &&
-        modalBrief && (
-          <div
-            className={
-              modalBg +
-              "transition-opacity duration-200 " +
-              (modalAppearing ? "ease-in opacity-100 " : "ease-out opacity-0 ")
-            }
-            onClick={() => window.history.back()}
-          >
-            <div className="absolute inset-6">
-              <Box
-                onClick={(e) => e.stopPropagation()}
-                className={
-                  "absolute inset-0 m-auto w-max h-max max-w-full max-h-full " +
-                  "p-6 overflow-x-clip overflow-y-auto " +
-                  "shadow-lg " +
-                  "transition-transform duration-200 origin-center " +
-                  (modalAppearing ? "ease-in scale-100 " : "ease-out scale-0 ")
-                }
-              >
-                <ShareBox
-                  cid={modalCId}
-                  brief={modalBrief}
-                  record={modalRecord}
-                  locale={locale}
-                  backButton={() => window.history.back()}
-                />
-              </Box>
-            </div>
-          </div>
-        )
-      }
+      modal={modal}
     >
-      <AccordionLike
-        hidden={showExclusiveMode !== null}
-        header={
-          <>
-            <span className="text-xl font-bold font-title">
-              {t("inputId")}:
-            </span>
-            <Input
-              className="ml-4 w-20"
-              actualValue=""
-              updateValue={gotoCId}
-              isValid={(t) => v.safeParse(CidSchema(), t).success}
-              left
-            />
-            <span className={cidFetching ? "inline-block " : "hidden "}>
-              <SlimeSVG />
-              Loading...
-            </span>
-            <span className="ml-1 inline-block">{cidErrorMsg}</span>
-          </>
-        }
-      >
-        <p className="pl-2 text-justify">{t("inputIdDesc")}</p>
-        <p className="pl-2 text-justify">
-          {t.rich("inputIdDesc2", {
-            url: () => <SmallDomainShare />,
-          })}
-        </p>
-        {process.env.NODE_ENV === "development" && (
-          <div className="mt-3 ml-2">
-            <h4 className="mb-1 ">
-              <span className="text-lg font-bold font-title">
-                {t("inputDirect")}:
-              </span>
-              <Input
-                className="ml-4 w-20"
-                actualValue=""
-                updateValue={(cid) =>
-                  window
-                    .open(`/${locale}/play?cid=${cid}&lvIndex=0`, "_blank")
-                    ?.focus()
-                }
-                isValid={(t) => v.safeParse(CidSchema(), t).success}
-                left
-              />
-            </h4>
-            <p className="pl-2 text-justify text-sm ">
-              ({t("inputDirectDevonly")})
-            </p>
-          </div>
-        )}
-      </AccordionLike>
-      <AccordionLike
-        hidden={showExclusiveMode !== null && showExclusiveMode !== "recent"}
-        expanded={showAllMode === "recent"}
-        reset={() => window.history.back()}
-        header={
-          <span className="text-xl font-bold font-title">
-            {t("recentPlay")}
-          </span>
-        }
-      >
-        <ChartList
-          recentBrief={recentBrief}
-          maxRow={chartListMaxRow}
-          fetchAdditional={() => setFetchRecentAll(true)}
-          creator
-          href={(cid) => `/share/${cid}`}
-          onClick={(cid) =>
-            openModal(cid, recentBrief?.find((b) => b.cid === cid)?.brief)
-          }
-          showLoading
-          additionalOpen={showAllMode === "recent"}
-          setAdditionalOpen={(open) => {
-            if (open) {
-              goExclusiveMode("recent");
-            } else {
-              window.history.back();
-            }
-          }}
-        />
-      </AccordionLike>
       <AccordionLike
         hidden={showExclusiveMode !== null && showExclusiveMode !== "popular"}
         expanded={showAllMode === "popular"}
