+++ i18n/ja/main.js
@@ -18,6 +18,14 @@ export default {
     updating: "更新をダウンロード中...",
     updateDone: "更新が完了しました！",
     updateFailed: "更新に失敗しました...",
+    inputId: "譜面IDを入力",
+    inputIdDesc:
+      "プレイしたい譜面の ID を知っている場合はこちらに入力してください。",
+    inputIdDesc2:
+      "譜面のURL (<url></url>) にアクセスすることでもプレイできます。",
+    inputDirect: "譜面IDを指定し直接プレイ画面に飛ぶ",
+    inputDirectDevonly: "dev環境限定、 /share/cid のパスが使えない代わり",
+    recentPlay: "最近プレイした譜面",
     about: {
       title: "Falling Nikochan とは？",
       titleShort: "概要",
@@ -25,14 +33,6 @@ export default {
     play: {
       title: "プレイする",
       titleShort: "プレイ",
-      inputId: "譜面IDを入力",
-      inputIdDesc:
-        "プレイしたい譜面の ID を知っている場合はこちらに入力してください。",
-      inputIdDesc2:
-        "譜面のURL (<url></url>) にアクセスすることでもプレイできます。",
-      inputDirect: "譜面IDを指定し直接プレイ画面に飛ぶ",
-      inputDirectDevonly: "dev環境限定、 /share/cid のパスが使えない代わり",
-      recentPlay: "最近プレイした譜面",
       popular: "人気の譜面",
       popularDesc: "直近 {popularDays} 日間にプレイされた回数の多い譜面です。",
       latest: "新着譜面",
